pulse-review-scraper/requirements.txt
requests>=2.31.0
beautifulsoup4>=4.12.2
python-dateutil>=2.9.0
lxml>=5.2.0
tqdm>=4.66.0
